chore: factor out progress-bar rendering (AI)
